+++ src/main/java/net/hecco/bountifulfares/compat/jei/PropagationRecipe.java
@@ -1,0 +1,6 @@
+package net.hecco.bountifulfares.compat.jei;
+
+import net.minecraft.item.Item;
+
+public class PropagationRecipe {
+}