+++ src/main/java/net/hecco/bountifulfares/compat/jei/category/FermentingRecipeCategory.java
@@ -25,9 +25,9 @@ import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
+@SuppressWarnings("removal")
 public class FermentingRecipeCategory implements IRecipeCategory<FermentationRecipe> {
     private final IDrawable fermentationVessel;
     private final IDrawable background;